refactor : cleansing package and name
