+++ src/main/java/com/geupjo/koreantiger/dto/StudentRankingDto.java
@@ -1,4 +1,4 @@
-package com.geupjo.koreantiger.dto.response;
+package com.geupjo.koreantiger.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
