+++ src/main/java/com/geupjo/koreantiger/dto/response/RankingBoardResponseDto.java
@@ -1,10 +1,11 @@
 package com.geupjo.koreantiger.dto.response;
 
+import com.geupjo.koreantiger.dto.StudentRankingDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.ArrayList;
 
-public record RankingBoardDto(
+public record RankingBoardResponseDto(
         @Schema(description = "우리학교랭킹 50위까지")
         ArrayList<StudentRankingDto> SchoolRanking,
