+++ src/main/java/com/geupjo/koreantiger/dto/response/StudentProfileResponseDto.java
@@ -4,7 +4,7 @@ import com.geupjo.koreantiger.enums.StudentProfileTitle;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "학생 프로필 응답DTO 입니다")
-public record StudentProfileDto(
+public record StudentProfileResponseDto(
         @Schema(description = "학생이름")
         String name,
