--- src/main/java/com/geupjo/koreantiger/service/Service.java
@@ -1,4 +1,0 @@
-package com.geupjo.koreantiger.service;
-
-public class Service {
-}
