--- src/main/java/com/geupjo/koreantiger/service/facade/Facade.java
@@ -1,4 +1,0 @@
-package com.geupjo.koreantiger.service.facade;
-
-public class Facade {
-}
